Don't chmod file if already executable
If the installation of the plugin is not owned by the current user, then
it's unlikely that they will have permission to change the mode of the
bashdb file.

To protect against that, we avoid chmod'ing the bundled bashdb if it's
already got the execute bit set, allowing this to be done once for the
installation by a user with write permission.